recc-engine/pyproject.toml
[project]
name = "recc-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=1.4.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "tmdbsimple>=2.9.1",
]
